Moving Register drawing functions to dedicated class
