+++ pulser/register/register.py
@@ -24,9 +24,10 @@ from numpy.typing import ArrayLike
 
 import pulser
 from pulser.register.base_register import BaseRegister
+from pulser.register._reg_drawer import RegDrawer
 
 
-class Register(BaseRegister):
+class Register(BaseRegister, RegDrawer):
     """A 2D quantum register containing a set of qubits.
 
     Args:
